requirements.txt
fastapi
uvicorn
motor
pydantic
langchain
langgraph
langchain-google-genai
python-dotenv
